Move JSONValue to http.d.ts
