+++ apps/astro-app/src/utils/data/deserialization.ts
@@ -1,5 +1,5 @@
 import type { Datatype } from "@/types/data";
-import type { JSONValue } from "@utils/http";
+import type { JSONValue } from "@/types/http";
 
 /**
  * Attempts to parse a database value, expecting the given datatype. Throws an error if unsuccessful.
